app/requirements.txt
cassandra-driver
venv-pack
pathvalidate
